requirements.txt
Flask==3.1.0
pymongo==4.5.0
python-dotenv==1.0.0
user-agents==2.2.0